requirements.txt
Django>=3.2.2,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.8.6,<2.9.0
flake8>=3.9.2,<3.10.0
django-filter>=2.4.0, <2.5.0
django-graphql-jwt
pyyaml
uritemplate
coreapi
django-cors-headers